+++ R/raw.getType.R
@@ -1,4 +1,6 @@
 #' Type of rawData item
+#' @param rawData type
+#' @param ID uique ID
 #' @returns string with type, such as "XRD", "AFM", etc.
 #' @export
 raw.getType <- function(rawData, ID) {